refactor: streamline DTO to VO conversion and update user login return type
